[GHO-75] Refactor the formatter for the related articles so that the main logic can be used elsewhere
