+++ html/modules/custom/gho_fields/config/schema/gho_fields.schema.yml
@@ -19,6 +19,11 @@ field.formatter.settings.gho_number:
     precision:
       type: integer
       label: 'Precision'
+field.formatter.settings.gho_related_articles:
+  type: mapping
+  label: 'GHO related articles formatter settings'
+  mapping:
+    # No settings.
 
 # Widgets.
 field.widget.settings.gho_dataset_link:
@@ -29,3 +34,8 @@ field.widget.settings.gho_further_reading_link:
   # Inherit the settings from the link widget.
   type: field.widget.settings.link_default
   label: 'GHO further reading link widget settings'
+field.widget.settings.gho_menu_select:
+  type: mapping
+  label: 'GHO menu select widget settings'
+  mapping:
+    # No settings.
